feat(procedure and util): Incomplete implementation of custom multiframe processing
Custom multiframe processing consists of separating each frame to its
own encoding process (for avif and webp), but treating it at a whole, by
averaging the metric results of all frames. This, in fact, is the
feature that is intended to be complete at the end of the current
branch.

It is incomplete: the pngs stapling process - decoded pngs
joining into a single apng, has a bug.