refactor(Procedure's quality settings): Introduced constants for easy tweaking the quality settings of compressions
In this case only the low end threshold have variables, since they will
need to be tweaked.
They need tweaking because the general experiment is
to be done using the near lossless level of compression, thus a
threshold defining the frontier between lossy and near lossless needs to
be found (for each format).