Update readme with software state (#18)
* Update with upstream changes (#7)

* Restructure and additions

Additions:
 * Dataset set up
 * Quick introduction

* Fixed ToC reference to hardware reqs

* Tweaks on the readme

* [Refactor] Use jpeg extended; Multi-frame images are now supported

 * Jpeg extended allows compression of images with bit depth greater than 8

* Updated requirements.txt

* Python generated documentation for github pages

* [Dicom Parser] 16 bit images supported; Multi-frame supported

* Multi-frame images extracted from dicom are saved in tiff format - thus, can only now be read from cwebp
* Minor cleanup

* Chenged scripts path

* Multi-frame images are now also encoded to apng

 * That is because cjxl doesn't support .tiff as input, but supports .apng (however, cavif-rs only supports jpg/png, thus I, for now, won't be evaluating multi-frame for avif)
 * Also, the program dicom_parser.py deletes all files from images/dataset before writing them to that directory - because I would need to manually delete those images whenever I need to re-run the script over the same images

* Command typo

* [Fix] Wrong unzip directory output

* Functional pipeline supporting gray images (#1)

* chore(requirements.txt): Update libraries and added necessary dependencies

* feat(dicom-parser.py): Add number of frames of the image on the lossless output filename

BREAKING CHANGE: Information on the lossless file name is changed (one
more attribute)

* feat(Full project): Evaluation for multiple frame images (jxl)

* fix(Full project): AVIF and WEBP output transcoding from rgb to gray (when appropriate)

Also:
    - Removed multi-frame TIFFs from webp encoding coverage
    - Added TODOs
    - Script now removes junk from previous runs

* Command typo

* [Fix] Wrong unzip directory output

* Custom multiframe support for avif and webp (#2)

* feat(procedure and util): Incomplete implementation of custom multiframe processing

Custom multiframe processing consists of separating each frame to its
own encoding process (for avif and webp), but treating it at a whole, by
averaging the metric results of all frames. This, in fact, is the
feature that is intended to be complete at the end of the current
branch.

It is incomplete: the pngs stapling process - decoded pngs
joining into a single apng, has a bug.

* feat(procedure and custom_apng): Custom multiframe processing complete for avif

* feat(procedure and util): Custom multiframe processing complete for webp

* Graphs visually representing procedure results (#3)

* feat(visualize_data and dicom_parser): Line graphs for metric values comparizon is fully parametrizable

We can parametrize which format, modality and metric that we want to
evaluate
And replicated changes from #f9ce292

* feat(visualize_data): Bar graphs for metric values comparizon per quality setting

* Implement support for colorized medical images (#4)

* fix(dicom_parser): Body part tag is no longer required (is now optional)

Since it is optional (3) in dicom files - as stated by the
dicom
standard

* Graphs visually representing procedure results (#3)

* feat(visualize_data and dicom_parser): Line graphs for metric values comparizon is fully parametrizable

We can parametrize which format, modality and metric that we want to
evaluate
And replicated changes from #f9ce292

* feat(visualize_data): Bar graphs for metric values comparizon per quality setting

* feat(dicom_parser and metrics): Support for colored images for pre-processing

And some support from the similarity metrics (incomplete)

* feat(Full project): Support for colored images for full pipeline

Note: mse == 0.0 => psnr == inf, the latter appearing as value "0"
in
graphs.

* chore(Idea Project Files): Added project configuration files for

* Results segmentation (#5)

* feat(Main pipeline runnables): Sub-divided results into depth/spp/bps filters

For the post-processing (visualization) the filters are required
parameters.
Unrelated minimal changes for the pre-processing module.

* feat(Results visualization and metrics): Wildcard search and bugfixes

Metric router only called mse when image is multiframe. Also,
added
kwargs in that situation (since it could be any metric, kwargs may
be
needed)

* feat(Results visualization - bar graph): Dynamic y-window and wildcard error fix

The y-window sizing is based on the difference (a percentage of it)
between the highest and lowest bars - give a small space above the
highest bar and below the lowest one.
Wildcard error - fix was to use
list.extend to append multiple values instead of list.append.

* chore(Results visualization - bar graph): Removed done TODO tasks

* refactor(Procedure's quality settings): Introduced constants for easy tweaking the quality settings of compressions

In this case only the low end threshold have variables, since they will
need to be tweaked.
They need tweaking because the general experiment is
to be done using the near lossless level of compression, thus a
threshold defining the frontier between lossy and near lossless needs to
be found (for each format).

* chore(Github workflow script): Simplify wiki github actions into basic workflow

* chore(Github workflow script): Removed gh actions yaml file

* feat(Visualize data): Metrics per metrics graph (#6)

Allows evaluating the relationship between different metrics (such as
how does the compression ratio evolves along increasing ssim similarity
values).

* refactor(Main and post processing modules, jpeg secondary pipeline, project files, readme): Code cleanup

Extras:
 * Idea project files - script running configurations
 * Add
dependency notice on readme

* Polish the setting up part

* How to run and other minor sections

* Simplified instalation of the ic codecs